Move Combine and Traverse to a separate package
