+++ core/src/main/scala/io/scalac/mesmer/core/module/AkkaPersistenceModule.scala
@@ -2,10 +2,10 @@ package io.scalac.mesmer.core.module
 import com.typesafe.config.{ Config => TypesafeConfig }
 
 import io.scalac.mesmer.core.model.Version
-import io.scalac.mesmer.core.module.Module.Combine
 import io.scalac.mesmer.core.module.Module.CommonJars
 import io.scalac.mesmer.core.module.Module.JarsNames
-import io.scalac.mesmer.core.module.Module.Traverse
+import io.scalac.mesmer.core.typeclasses.Combine
+import io.scalac.mesmer.core.typeclasses.Traverse
 import io.scalac.mesmer.core.util.LibraryInfo.LibraryInfo
 
 sealed trait AkkaPersistenceMetricsModule extends MetricsModule {
