+++ core/src/main/scala/io/scalac/mesmer/core/typeclasses/Combine.scala
@@ -1,0 +1,5 @@
+package io.scalac.mesmer.core.typeclasses
+
+trait Combine[T] {
+  def combine(first: T, second: T): T
+}
