+++ core/src/main/scala/io/scalac/mesmer/core/typeclasses/Traverse.scala
@@ -1,0 +1,5 @@
+package io.scalac.mesmer.core.typeclasses
+
+trait Traverse[F[_]] {
+  def sequence[T](obj: F[T]): Seq[T]
+}
